refactor: 오타 수정 및 undo 기능 수정
