+++ webpack.dev.ts
@@ -1,8 +1,9 @@
 import path from 'path'
-import common from './webpack.common'
-import merge from 'webpack-merge'
 import { Configuration as WebpackConfiguration } from 'webpack'
 import { Configuration as WebpackDevServerConfiguration } from 'webpack-dev-server'
+import merge from 'webpack-merge'
+
+import common from './webpack.common'
 
 interface Configuration extends WebpackConfiguration {
   devServer?: WebpackDevServerConfiguration
